[FIx] 몬스터 서치 수정
하이어라키에 활성화 된 오브젝트만 참조. 안개 수정. 유저 재화 강화값 추가.

+++ Assets/Resources/Scripts/Manager/InGameManager.cs
@@ -160,13 +160,13 @@ public class InGameManager : MonoBehaviour
         var tempgold = playerData._defaultGold + playerData.totalGoldEnhance;
 
         Debug.Log("시작 시 기본 재화 + 강화값 " + tempgold);
-        Debug.Log("시작 시 기본 재화 " + playerData._defaultGold);
+        Debug.Log("시작 시 기본 재화 " + playerData._defaultGold); 
         Debug.Log("시작 시 강화값 " + playerData.totalGoldEnhance);
 
         StartCoroutine(UpdateGetGold(tempgold));
     }
 
-    public IEnumerator UpdateGetGold(int getgoldvalue)
+    public IEnumerator UpdateGetGold(float getgoldvalue)
     {
         while (true)
         {
